[UPDATE] #BLOCK MANAGEMENT# Refactoring, correcting c-lang errors.

+++ size_control.c
@@ -6,7 +6,7 @@
 
 size_t align(size_t sz) {
     for (size_t i = 0; i != 32; ++i)
-        if (1 << i > sz)
-            return 1 << i;
+        if ((size_t)1 << i > sz)
+            return (size_t)1 << i;
     return sz;
 }
